Initial upload - LC Callbacks.livecode
Initial upload of Peter Bogdanoff's LiveCode Video Callbacks example file - LC Callbacks.livecode

This file uses platform specific, wmv and mov, files to provide timed feedback to a user based on their progress through a video file. Logic is incuded to determine the current platform of Mac or Windows and return the specific video file. Within these limitations the video player object in this project will display a message to the user based on their progress viewing a video. This example could be customized to send xAPI statements as part of the callback function, but this functionality is not included in the initial release.